+++ src/main/java/umc/cockple/demo/domain/exercise/converter/command/ExerciseGuestCommandMapper.java
@@ -3,7 +3,7 @@ package umc.cockple.demo.domain.exercise.converter.command;
 import org.springframework.stereotype.Component;
 import umc.cockple.demo.domain.exercise.domain.Exercise;
 import umc.cockple.demo.domain.exercise.domain.Guest;
-import umc.cockple.demo.domain.exercise.dto.ExerciseCancelDTO;
+import umc.cockple.demo.domain.exercise.dto.participation.ExerciseCancelDTO;
 import umc.cockple.demo.domain.exercise.dto.ExerciseGuestInviteDTO;
 
 @Component
